// frontend/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  alerts: typeof alerts;
  api_data_sources: typeof api_data_sources;
  chat: typeof chat;
  chat_history: typeof chat_history;
  columns: typeof columns;
  export_data: typeof export_data;
  http: typeof http;
  http_actions: typeof http_actions;
  http_valdation_errors: typeof http_valdation_errors;
  jobs: typeof jobs;
  logs: typeof logs;
  middleware: typeof middleware;
  project_groupings: typeof project_groupings;
  projects: typeof projects;
  prompt: typeof prompt;
  relationships: typeof relationships;
  rows: typeof rows;
  scheduled_actions: typeof scheduled_actions;
  search_engines: typeof search_engines;
  service_credentials: typeof service_credentials;
  sheets: typeof sheets;
  system_settings: typeof system_settings;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;